- Se cambio la generacion de IDs random por generacion incremental. - Se hicieron los siguientes endpoints para Materia: POST, PUT, DELETE, GET (by ID) y GET (by Nombre).

+++ src/main/java/ar/edu/utn/frbb/tup/business/AlumnoService.java
@@ -6,7 +6,7 @@ import ar.edu.utn.frbb.tup.model.exception.CorrelatividadesNoAprobadasException;
 import ar.edu.utn.frbb.tup.model.exception.EstadoIncorrectoException;
 
 public interface AlumnoService {
-    void aprobarAsignatura(int materiaId, int nota, long dni) throws EstadoIncorrectoException, CorrelatividadesNoAprobadasException;
+    void aprobarAsignatura(long materiaId, int nota, long dni) throws EstadoIncorrectoException, CorrelatividadesNoAprobadasException;
 
     Alumno crearAlumno(AlumnoDto alumno);
