+++ src/main/java/ar/edu/utn/frbb/tup/business/AsignaturaService.java
@@ -3,7 +3,7 @@ package ar.edu.utn.frbb.tup.business;
 import ar.edu.utn.frbb.tup.model.Asignatura;
 
 public interface AsignaturaService {
-    Asignatura getAsignatura(int materiaId, long dni);
+    Asignatura getAsignatura(long materiaId, long dni);
 
     void actualizarAsignatura(Asignatura a);
 }
